Get all statuses for my signers
There's a bit of craziness in here: it shells out to gnupg to get the
signatures. This is because fetching the list of signatures is
deprecated in GPGME, and therefore not enabled in the Ruby bindings.
When I asked in #gnupg on Freenode I was instructed to shell out
instead.

This fetches the signing keys in the Web server process. That's how you
know that this is a prototype.